+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Feature: [#3087] Allow replacing existing station elements directly in the station construction tool.
 - Change: [#3095] The Tile Inspector can now be accessed in the Scenario Editor when the cheats menu is enabled.
 - Change: [#3098] Automatically detect Steam installation of Locomotion.
+- Fix: [#2689] Title screen music not playing after switching audio device.
 - Fix: [#3092] Drag selection area is not correctly removed after dragging station elements.
 - Fix: [#3096] Buildings are removed over time in the scenario editor (original bug).